pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ah-premium-lab"
version = "0.1.0"
description = "Research toolkit for A/H premium dashboard, statistical tests, and cost sensitivity reporting"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "AH Premium Lab" }]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "scipy>=1.11",
  "pyyaml>=6.0",
  "pydantic>=2.8",
  "yfinance>=0.2.61",
  "statsmodels>=0.14",
  "streamlit>=1.42",
  "plotly>=5.24",
  "jinja2>=3.1",
  "pytest>=8.3",
]

[project.optional-dependencies]
dev = [
  "black>=24.10",
  "isort>=5.13",
  "pre-commit>=4.0",
  "ruff>=0.9",
  "mypy>=1.15",
  "types-PyYAML>=6.0.12.20241230",
]

[project.scripts]
ah-premium = "ah_premium_lab.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ah_premium_lab = ["report/templates/*.html.j2"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "D"]
ignore = ["D202", "D203", "D212"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["ah_premium_lab"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
mypy_path = "src"
